Pull in goisilon ownership fix
The goisilon bindings incorrectly attempted to change ownership of a
directory after creation. This was redundant and also failed unless the
parent directory had an ACL that allowed the user the "std_write_owner"
right.
Also pull in verion of the library that supports changing ownership of a
file/directory.